The robot will execute the last command until STOP command received
The robot will keep executing the last command received until the STOP
command is received. This should make the use of
delay(MOVE_DURATION_MILLIS) unnecessary.
Also there will not be a continous stream of the same command but just
one. So the "while" loop to read from Serian needs to be converted in a
simple "if".